Apply Prettier
