+++ src/components/Form.tsx
@@ -1,9 +1,8 @@
-import { useForm } from 'react-hook-form';
-import FormElement from 'components/FormElement';
-import { FormProps } from 'models/form';
+import { useForm } from "react-hook-form";
+import FormElement from "components/FormElement";
+import { FormProps } from "models/form";
 
-const Form = ({elements, onSubmit}: FormProps) => {
-  
+const Form = ({ elements, onSubmit }: FormProps) => {
   const {
     register,
     handleSubmit,
@@ -12,17 +11,10 @@ const Form = ({elements, onSubmit}: FormProps) => {
 
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
-      {
-        elements.map(element => <FormElement
-                                  key={element.id}
-                                  id={element.id}
-                                  name={element.name} 
-                                  option={element.option} 
-                                  register={register} 
-                                  errors={errors}
-                                />)
-      }
-      <button type="submit">등록</button>
+      {elements.map(element => (
+        <FormElement key={element.id} id={element.id} name={element.name} option={element.option} register={register} errors={errors} />
+      ))}
+      <button type='submit'>등록</button>
     </form>
   );
 };
