+++ src/components/Grid.tsx
@@ -1,16 +1,19 @@
-import 'ag-grid-community/styles/ag-grid.css';
-import 'ag-grid-community/styles/ag-theme-alpine.css';
-import { AgGridReact } from 'ag-grid-react';
-import { GridOptions } from 'ag-grid-community';
+import "ag-grid-community/styles/ag-grid.css";
+import "ag-grid-community/styles/ag-theme-alpine.css";
+import { AgGridReact } from "ag-grid-react";
+import { GridOptions } from "ag-grid-community";
 
 const Grid = ({ data, options }: { data: any; options: GridOptions }) => {
-  const gridOption: GridOptions = Object.assign({
-    onGridReady: () => {},
-    pagination: true,
-    paginationPageSize: 5,
-    paginationPageSizeSelector: false,
-    paginationAutoPageSize: false,
-  }, options);
+  const gridOption: GridOptions = Object.assign(
+    {
+      onGridReady: () => {},
+      pagination: true,
+      paginationPageSize: 5,
+      paginationPageSizeSelector: false,
+      paginationAutoPageSize: false,
+    },
+    options
+  );
 
   return (
     <div>
@@ -18,7 +21,7 @@ const Grid = ({ data, options }: { data: any; options: GridOptions }) => {
         columnDefs={data.columns}
         rowData={data.rows}
         className={"ag-theme-alpine"}
-        domLayout={'autoHeight'}
+        domLayout={"autoHeight"}
         onGridReady={gridOption.onGridReady}
         pagination={gridOption.pagination}
         paginationPageSize={gridOption.paginationPageSize}
