+++ src/models/form.ts
@@ -1,42 +1,42 @@
 import { FieldErrors, FieldValues, SubmitHandler, UseFormRegister } from "react-hook-form";
 
 export interface FormProps {
-  elements: IFormElement[],
-  onSubmit: SubmitHandler<FieldValues>,
+  elements: IFormElement[];
+  onSubmit: SubmitHandler<FieldValues>;
 }
 
 export interface IFormElement {
-  id: string,
-  name: string,
-  type?: string,
+  id: string;
+  name: string;
+  type?: string;
   option?: {
-    required?: string,
+    required?: string;
     pattern?: {
-      value: RegExp,
-      message: string,
-    },
+      value: RegExp;
+      message: string;
+    };
     minLength?: {
-      value: number,
-      message: string,
-    },
-  },
+      value: number;
+      message: string;
+    };
+  };
 }
 
 export interface FormElementProps {
-  id: string,
-  name: string,
-  type?: string,
+  id: string;
+  name: string;
+  type?: string;
   option?: {
-    required?: string,
+    required?: string;
     pattern?: {
-      value: RegExp,
-      message: string,
-    },
+      value: RegExp;
+      message: string;
+    };
     minLength?: {
-      value: number,
-      message: string,
-    },
-  },
-  register: UseFormRegister<any>,
-  errors: FieldErrors<FieldValues>,
+      value: number;
+      message: string;
+    };
+  };
+  register: UseFormRegister<any>;
+  errors: FieldErrors<FieldValues>;
 }
