+++ src/pages/GridPage.tsx
@@ -1,7 +1,7 @@
 import Grid from "components/Grid";
 import "ag-grid-community/styles/ag-grid.css";
 import "ag-grid-community/styles/ag-theme-alpine.css";
-import { retrieveCars } from 'services/grid';
+import { retrieveCars } from "mocks/grid";
 import useFetch from "hooks/cmn/useFetch";
 
 const GridPage = () => {
@@ -13,7 +13,7 @@ const GridPage = () => {
         <h2>Grid page</h2>
         <p>The list of cars</p>
       </div>
-      <Grid 
+      <Grid
         data={data}
         options={{
           pagination: true,
